+++ src/index.js
@@ -1,6 +1,6 @@
-import BasicAuthPlugin from './plugins/BasicAuthPlugin'
-import BasicAuthProxyTicketPlugin from './plugins/BasicAuthProxyTicketPlugin'
-import SessionIdPlugin from './plugins/SessionIdPlugin'
+import BasicAuthPlugin from './BasicAuthPlugin'
+import BasicAuthProxyTicketPlugin from './BasicAuthProxyTicketPlugin'
+import SessionIdPlugin from './SessionIdPlugin'
 
 export {
   BasicAuthPlugin,
